fix build with overlayed catkin workspace
catkin should be the first find_packaged component.
Usually it's no problem to do it any other way, but apparently
one of the intermediate ones can modify the state in such a way that the correct catkin
instance is skipped and the overlayed one is used...

As this change is not harmful for anything else but addresses an edge-case, I will merge it.